requirements.txt
# Core dependencies
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# MongoDB
pymongo==4.6.0
motor==3.3.1

# Data processing and ML
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2

# Device fingerprinting
user-agents==2.2.0
ua-parser==0.18.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1